Compile the web runtime against the bindings applications compile against
v0.1.96 was published with four `error[E0308]` in the web runtime:
`pointer_position(x: f64, y: f64)` called with the `i32` that
`event.offset_x()` returns. Every application's web build failed on it, while
this repository's own wasm job stayed green on the same source.

`--cfg web_sys_unstable_apis` is why. It is not an additive opt-in: `web-sys`
declares `MouseEvent::offset_x`/`offset_y` twice, `-> i32` under
`cfg(not(web_sys_unstable_apis))` and `-> f64` under `cfg(web_sys_unstable_apis)`.
The flag lived in this repository's cargo configuration, which applies to builds
started here and to nothing an application runs, so the framework was compiled
against one signature and every consumer against the other. Clippy ran under the
flag too and reported the `as f64` as an `unnecessary_cast`; removing it is what
broke the published crate.

Nothing needed the flag. Removing it from a clean checkout reproduces exactly
those four errors and no others, and `web_media` already states the rule it
violated: an application should not have to set a rustc flag to get a lock
screen, which is why the Media Session API is reached through `js_sys::Reflect`
rather than through bindings behind the flag. The flag is gone from both cargo
configurations and the casts are back.

Two things now hold that line. `no_cargo_config_enables_unstable_web_sys_bindings`
fails if any `.cargo/config.toml` enables it again. And the release canary builds
`apps/isolated-demo` for the web: it carries no cargo configuration of its own, so
it is the only tree in the repository that compiles a release the way an
application does. It was already built for desktop and Android there; the web,
the one target whose configuration can differ, was the one it skipped.

Verified on a copy of the isolated demo with no cargo configuration, resolving
the framework through `[patch.crates-io]`: it type-checks for wasm32 with the
casts and reproduces all four errors without them.

+++ crates/cranpose/src/web.rs
@@ -420,8 +420,8 @@ pub async fn run(
         let wheel_canvas = canvas.clone();
         let request_frame = request_frame.clone();
         let closure = Closure::wrap(Box::new(move |event: WheelEvent| {
-            let x = event.offset_x();
-            let y = event.offset_y();
+            let x = event.offset_x() as f64;
+            let y = event.offset_y() as f64;
             let logical = platform.borrow().pointer_position(x, y);
 
             // What the sample means -- zoom, rotary, axis-swapped scroll, or
@@ -458,8 +458,8 @@ pub async fn run(
         let request_frame = request_frame.clone();
         let closure = Closure::wrap(Box::new(move |event: PointerEvent| {
             event.prevent_default();
-            let x = event.offset_x();
-            let y = event.offset_y();
+            let x = event.offset_x() as f64;
+            let y = event.offset_y() as f64;
             let logical = platform.borrow().pointer_position(x, y);
             if let Ok(mut app_mut) = app.try_borrow_mut() {
                 app_mut.set_pointer_source(web_pointer_source(&event));
@@ -479,8 +479,8 @@ pub async fn run(
         let closure = Closure::wrap(Box::new(move |event: PointerEvent| {
             event.prevent_default();
             let _ = pointer_canvas.set_pointer_capture(event.pointer_id());
-            let x = event.offset_x();
-            let y = event.offset_y();
+            let x = event.offset_x() as f64;
+            let y = event.offset_y() as f64;
             let logical = platform.borrow().pointer_position(x, y);
             if let Ok(mut app_mut) = app.try_borrow_mut() {
                 app_mut.set_pointer_source(web_pointer_source(&event));
@@ -501,8 +501,8 @@ pub async fn run(
         let request_frame = request_frame.clone();
         let closure = Closure::wrap(Box::new(move |event: PointerEvent| {
             event.prevent_default();
-            let x = event.offset_x();
-            let y = event.offset_y();
+            let x = event.offset_x() as f64;
+            let y = event.offset_y() as f64;
             let logical = platform.borrow().pointer_position(x, y);
             if let Ok(mut app_mut) = app.try_borrow_mut() {
                 app_mut.set_pointer_source(web_pointer_source(&event));
